+++ pkg/apirouter/routes/v1/v1_router.go
@@ -3,6 +3,7 @@ package v1
 
 import (
 	"github.com/RoadTripMoustache/iris_api/pkg/apirouter/routes/v1/ideas"
+	"github.com/RoadTripMoustache/iris_api/pkg/apirouter/routes/v1/images"
 	"github.com/gorilla/mux"
 )
 
@@ -22,4 +23,5 @@ func New(router *mux.Router) *APIV1Router {
 // InitRoutes - Initialize all the routes for the "v1" path.
 func (a *APIV1Router) InitRoutes() {
 	ideas.New(a.MuxRouter, a.Path).InitRoutes()
+	images.New(a.MuxRouter, a.Path).InitRoutes()
 }