+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # CHANGELOG
 
 ## Unreleased
-
+### Added
+- feat: Add endpoint to upload and get a image